eliminate dead code, move everything else under edx_chat_bot
