--- integrations/slack/requirements.txt
@@ -1,2 +1,0 @@
-
-limbo==4.4.2
